+++ tests/Unit/Client/LogTest.php
@@ -10,7 +10,10 @@ use Spawnia\Sailor\Client\Log;
 class LogTest extends TestCase
 {
     const FILENAME = __DIR__.'/LogTest.log';
+
+    const QUERY = /** @lang GraphQL */ '{ foo }';
     const EXPECTED_JSON = /** @lang JSON */ '{"query":"{ foo }","variables":{"bar":42}}'."\n";
+    const VARIABLES = ['bar' => 42];
 
     protected function tearDown(): void
     {
@@ -26,14 +29,44 @@ class LogTest extends TestCase
         self::assertFileDoesNotExist(self::FILENAME);
 
         $log = new Log(self::FILENAME);
-        $log->request(/** @lang GraphQL */ '{ foo }', (object) ['bar' => 42]);
+        $log->request(/** @lang GraphQL */ self::QUERY, (object) self::VARIABLES);
 
+        self::assertFileExists(self::FILENAME);
         $contents = \Safe\file_get_contents(self::FILENAME);
         self::assertSame(self::EXPECTED_JSON, $contents);
 
-        $log->request(/** @lang GraphQL */ '{ foo }', (object) ['bar' => 42]);
+        $log->request(/** @lang GraphQL */ self::QUERY, (object) self::VARIABLES);
 
         $contents = \Safe\file_get_contents(self::FILENAME);
         self::assertSame(self::EXPECTED_JSON.self::EXPECTED_JSON, $contents);
+    }
+
+    public function testRequests(): void
+    {
+        $log = new Log(self::FILENAME);
+        $log->request(/** @lang GraphQL */ self::QUERY, (object) self::VARIABLES);
+        $log->request(/** @lang GraphQL */ self::QUERY, null);
+
+        $decoded = iterator_to_array($log->requests());
+        self::assertSame([
+            [
+                'query' => self::QUERY,
+                'variables' => self::VARIABLES,
+            ],
+            [
+                'query' => self::QUERY,
+                'variables' => null,
+            ],
+        ], $decoded);
+    }
+
+    public function testClear(): void
+    {
+        \Safe\file_put_contents(self::FILENAME, 'foo');
+
+        $log = new Log(self::FILENAME);
+        $log->clear();
+
+        self::assertFileDoesNotExist(self::FILENAME);
     }
 }