+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.7.0
+
+### Added
+
+- Add methods `Log::requests()` and `Log::clear()` to ease assertions on Log client https://github.com/spawnia/sailor/pull/25
+
 ## v0.6.0
 
 ### Added